requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==41.0.7
cssselect2==0.7.0
daphne==4.0.0
defusedxml==0.7.1
distro==1.9.0
Django==5.0
django-allauth==0.59.0
django-progressbarupload==0.1.7
django-redis==5.4.0
django-weasyprint==2.2.2
exceptiongroup==1.2.0
fonttools==4.47.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
joblib==1.3.2
kombu==5.3.5
lxml==5.1.0
msgpack==1.0.7
nltk==3.8.1
oauthlib==3.2.2
openai==1.10.0
packaging==23.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydyf==0.8.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
PyPDF2==3.0.1
pypdfium2==4.25.0
pyphen==0.14.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
social-auth-core==4.5.1
sqlparse==0.4.4
stripe==8.3.0
tinycss2==1.2.1
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==60.2
webencodings==0.5.1
zope.interface==6.1
zopfli==0.2.3
